pyproject.toml
[project]
name = "alpi-transcript"
version = "0.1.0"
description = "A software that transcribes insurance brokers' speech and generates a report and actionable items such as meetings and new insurance policies"
authors = [
    { name = "mzgolliph", email = "155550362+mzgolliph@users.noreply.github.com" }
]
readme = "README.md"

[tool.poetry]
packages = [
    { include = "app" },
    { include = "processing" }
]

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
google-cloud-speech = "^2.32.0"
ipykernel = "^6.29.5"
streamlit-webrtc = "^0.62.4"
av = "^14.3.0"
pydub = "^0.25.1"
pyaudio = "^0.2.14"
gradio = "^5.26.0"
dash = "^3.0.4"
python-dotenv = "^1.1.0"
dash-core-components = "^2.0.0"
dash-html-components = "^2.0.0"
setuptools = "^80.0.0"
wheel = "^0.45.1"
#openai-whisper = { git = "https://github.com/openai/whisper.git", rev = "v20231117" }
watchdog = "^6.0.0"
streamlit = "^1.44.1"
torch = "^2.7.0"
openai = "^1.76.0"
dotenv = "^0.9.9"

[tool.poetry.group.dev.dependencies]
streamlit = "^1.44.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
src = [".", "processing"]
extend-exclude = ["**/migrations/*"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "F",  # Pyflakes
    "UP", # pyupgrade
    "B",  # flake8-bugbear
    "SIM", # flake8-simplify
    "I",  # isort
]
